"Removed" LCM feature
Crates.io won't let me publish a package that depends on a non-Crates
crate. Commented it out in Cargo.toml as (1) I doubt that it will be
used, even relative to how little this will be used and (2) this way
should be easy to keep it maintained until LCM is on Crates.