Action node now uses `Result`
I decided that it actually makes a lot of sense for the condition node
to return a boolean - that is literally their purpose. I also imagine
that somehow having a `Result` in a condition would be difficult to do
in a correctly written condition node.

Action nodes, on the other hand, make perfect sense to have return a
`Result`.